ghost-os/apps/agent-worker/requirements.txt
patchright>=0.5
google-genai>=1.0
openai>=1.0
python-dotenv>=1.0
Pillow>=10.0
mss>=9.0
rich>=13.0
fastapi>=0.100.0
uvicorn>=0.20.0
psycopg2-binary>=2.9.0
